\name{MS1_pos}
\alias{MS1_pos}
\docType{data}
\title{
Example MS1 data set for id functions in ESI+
}
\description{
Example data set for id functions
}
\usage{data("MS1_pos")}
\format{
  List with 2 data frames
  \describe{
    \item{\code{peaklist}}{Data frame with 4 columns: m.z, RT, int and peakID.}
    \item{\code{rawScans}}{Data frame with 5 columns: m.z, RT, int, peakID and Scan.}
  }
}
\keyword{datasets}
