\name{MSMS1_neg}
\alias{MSMS1_neg}
\docType{data}
\title{
Example MS2 data set for id functions in ESI- (low energy)
}
\description{
Example data set for id functions
}
\usage{data("MSMS1_neg")}
\format{
  List with 2 data frames
  \describe{
    \item{\code{peaklist}}{Data frame with 4 columns: m.z, RT, int and peakID.}
    \item{\code{rawScans}}{Data frame with 5 columns: m.z, RT, int, peakID and Scan.}
  }
}
\keyword{datasets}
