\name{MSMS2_pos}
\alias{MSMS2_pos}
\docType{data}
\title{
Example MS2 data set for id functions in ESI+ (high energy)
}
\description{
Example data set for id functions
}
\usage{data("MSMS2_pos")}
\format{
  List with 2 data frames
  \describe{
    \item{\code{peaklist}}{Data frame with 4 columns: m.z, RT, int and peakID.}
    \item{\code{rawScans}}{Data frame with 5 columns: m.z, RT, int, peakID and Scan.}
  }
}
\keyword{datasets}
